--- my-app/server/models/Locations.js
@@ -1,9 +1,0 @@
-const { Schema, model } = require('mongoose');
-
-const locationSchema = new Schema({
-
-});
-
-const Location = model('Location', locationSchema);
-
-module.exports = Location;
