--- my-app/server/models/Matchup.js
@@ -1,24 +1,0 @@
-const { Schema, model } = require('mongoose');
-
-const matchupSchema = new Schema({
-  tech1: {
-    type: String,
-    required: true,
-  },
-  tech2: {
-    type: String,
-    required: true,
-  },
-  tech1_votes: {
-    type: Number,
-    default: 0,
-  },
-  tech2_votes: {
-    type: Number,
-    default: 0,
-  },
-});
-
-const Matchup = model('Matchup', matchupSchema);
-
-module.exports = Matchup;
